+++ src/Purebred/Storage/Notmuch.hs
@@ -70,6 +70,7 @@ import Control.Lens (view, over, set, firstOf, folded, Lens')
 import qualified Notmuch
 
 import Purebred.Types
+import Purebred.Storage.Tags
 import Purebred.System.Process (readProcess, proc)
 import Purebred.System (tryIO)
 import Purebred.Types.Error
