+++ test/TestTagParser.hs
@@ -23,7 +23,7 @@ import Data.List (isInfixOf)
 
 import qualified Data.Text as T
 import Purebred.Storage.Tags
-import Purebred.Types (TagOp(..), UserMessage(..), Name(..), MessageSeverity(..))
+import Purebred.Types (UserMessage(..), Name(..), MessageSeverity(..))
 
 import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.HUnit ((@=?), testCase, assertBool, assertFailure)
